Replace int_fast8_t with short

+++ ising.h
@@ -1,6 +1,6 @@
 #include "params.h"
 #include <stdint.h>
 
-void update(const float temp, int_fast8_t grid[L][L]);
+void update(const float temp, short grid[L][L]);
 
-double calculate(int_fast8_t grid[L][L], int *M_max);
+double calculate(short grid[L][L], int *M_max);
